requirements.txt
requests==2.32.3
configparser==7.0.0
drawpyo==0.1.3
gitpython==3.1.43
Flask==3.0.3
python-dotenv==1.0.1